Refactor tests into one file
